Add AUR installation method
AUR is Arch User Repository, "community-driven repository for Arch users. It contains package descriptions (PKGBUILDs) that allow you to compile a package from source with makepkg and then install it via pacman." 
I've made an AUR package that will allow any arch-based distro user to install and update microban to the latest version without manual version cheking&compiling. It will be easier for them to find it if you add information about this in the readme